Rearrange and cleanup help text
