+++ src/lib.rs
@@ -1,2 +1,3 @@
+pub mod cli;
 pub mod interval;
 pub mod password;
